+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ARIA_S1_GUNW` job type to hyp3-edc-uat deployment.
 
 ### Changed
+- The reserved `bucket_prefix` job spec parameter has been renamed to `job_id` and can be referenced as `Ref::job_id` within each step's `command` field.
 - The `job_id` parameter of the `ARIA_RAIDER` job type has been renamed to `gunw_job_id`.
 - The `AUTORIFT_ITS_LIVE` job type now accepts Sentinel-1 burst products.
 - The `OriginAccessIdentityId` has been renamed to `BucketReadPricipals` and now accepts multiple values.